Renames single and multiple item codecs
