+++ packages/xforms-engine/src/lib/codecs/items/SingleValueItemCodec.ts
@@ -1,7 +1,7 @@
 import type { SharedValueCodec } from '../getSharedValueCodec.ts';
 import { type CodecDecoder, type CodecEncoder } from '../ValueCodec.ts';
-import { BaseItemCodec } from '../BaseItemCodec.ts';
-import type { BaseItemCollectionCodec } from '../BaseItemCollectionCodec.ts';
+import { BaseItemCodec } from './BaseItemCodec.ts';
+import type { MultipleValueItemCodec } from './MultipleValueItemCodec.ts';
 
 // prettier-ignore
 export type SingleValueSelectRuntimeValues =
@@ -41,7 +41,7 @@ const encodeValueFactory = (
  * Value codec implementation for `<select1>` controls.
  *
  * Note: this implementation is a specialization of the same principles
- * underlying {@link BaseItemCollectionCodec}. It is implemented separately:
+ * underlying {@link MultipleValueItemCodec}. It is implemented separately:
  *
  * 1. to address a semantic difference between `<select>` and `<select1>`
  *    values: the former are serialized as a space-separated list, but that does
@@ -50,7 +50,7 @@ const encodeValueFactory = (
  * 2. as an optimization, as the more general implementation performs poorly on
  *    forms which we monitor for performance.
  */
-export class SingleValueSelectCodec extends BaseItemCodec<SingleValueSelectCodecValues> {
+export class SingleValueItemCodec extends BaseItemCodec<SingleValueSelectCodecValues> {
 	constructor(baseCodec: SharedValueCodec<'string'>) {
 		const encodeValue = encodeValueFactory(baseCodec);
